% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_county_ACAG.R
\name{get_census_geo}
\alias{get_census_geo}
\title{Helper function to grab and transform all census tracts within a given county
and initialize particulate matter column to be passed to get_census_pm}
\usage{
get_census_geo(cty_row, st)
}
\arguments{
\item{cty_row}{data frame - single county with geometry}

\item{st}{string - state that county resides in}
}
\value{
data frame of tract level data for county and state with empty PM column
}
\description{
Helper function to grab and transform all census tracts within a given county
and initialize particulate matter column to be passed to get_census_pm
}
