% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_county_ACAG.R
\name{get_census_pm}
\alias{get_census_pm}
\title{Helper function to compute mean area weighted PM2.5 concentration for a given
census tract}
\usage{
get_census_pm(census)
}
\arguments{
\item{census}{data frame - single census tract with geometry}
}
\value{
data frame with mean weight PM2.5 estimates propagated for that CT
}
\description{
Helper function to compute mean area weighted PM2.5 concentration for a given
census tract
}
