% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_county_ACAG.R
\name{get_county_geo}
\alias{get_county_geo}
\title{Helper function to compute PM2.5 concentrations for every census tract for each
county in this state}
\usage{
get_county_geo(st, cty_df, dal)
}
\arguments{
\item{st}{string - state}

\item{cty_df}{dataframe}
}
\description{
Helper function to compute PM2.5 concentrations for every census tract for each
county in this state
}
