% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_GEOID.R
\name{lookup_GEOID}
\alias{lookup_GEOID}
\title{Look up GEOID of geographies}
\usage{
lookup_GEOID(state = NULL, county = NULL, tract = NULL)
}
\arguments{
\item{state, }{character string of selected state. Full name or abbreviation. In the format "state".}

\item{county, }{character string of selected county. Name with or without parish/county/borough classification. In the format "county, state".}

\item{tract, }{character string of selected tract. Full name or just tract code. In the format "tract, county, state".}
}
\value{
Character string representing GEOID corresponding to user input.
}
\description{
Provides the user with the ability to look up GEOIDs for input geography.
}
\examples{
state_GEOID <- lookup_GEOID(state = "AL")
county_GEOID <- lookup_GEOID(county = "Abbeville County, SC")
tract_GEOID <- lookup_GEOID(tract = "Census Tract 9502, Abbeville County, SC")
multiple_GEOID <- lapply(c("AL", "AZ", "AR"), function(st){
                      lookup_GEOID(state = st)})
}
