% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pull_census_ACAG}
\alias{pull_census_ACAG}
\title{Pulls PM data at census level, either internally or externally}
\usage{
pull_census_ACAG(year)
}
\arguments{
\item{year}{}
}
\value{
dataframe
}
\description{
Pulls PM data at census level, either internally or externally
}
