% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_city_ACAG.R
\name{pull_city_ACAG}
\alias{pull_city_ACAG}
\title{City level particulate matter data}
\usage{
pull_city_ACAG(year, city_state = c())
}
\arguments{
\item{year, }{numeric object representing a selected year}

\item{city_state, }{character vector of selected cities}
}
\value{
Dataframe object broken down by census tracts in each city with
mean area-weighted PM2.5 values. If the city_state field is empty, PM2.5
data for census tracts in all cities is returned. If input year is
unavailable, returns an error.
}
\description{
Pulls PM2.5 data at a city level, either internally or externally. Years
2015 through 2018 are pre-available within the package, but year ____ through
2014 are available as an external pull.
}
\examples{
pull_city_ACAG(year = 2016, state = c("Akron_OH", "Albany_NY", "Altoona_PA"))
pull_city_ACAG(year = 2016)
}
