% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_county_ACAG.R
\name{pull_county_ACAG}
\alias{pull_county_ACAG}
\title{County level particulate matter data}
\usage{
pull_county_ACAG(year, state = c())
}
\arguments{
\item{year, }{numeric object representing a selected year}

\item{state, }{character vector of selected states}
}
\value{
Dataframe object broken down by counties in selected state with mean
area-weighted PM2.5 values. If the state field is empty, PM2.5 data for
counties in all states is returned. If input year is unavailable, returns an
error.
}
\description{
Pulls PM2.5 data at a county level, either internally or as a new pull. Years
2015 through 2018 are pre-available within the package, but years 2014 and
earlier are available as a pull combining PM2.5 raster files and tigris
shape files.
}
\examples{
acag_pm_dat_county <- pull_county_ACAG(year = 2016, state = c("AL", "AK", "AZ"))
acag_pm_dat_county <- pull_county_ACAG(year = 2016)
}
