% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_national_ACAG.R
\name{pull_national_ACAG}
\alias{pull_national_ACAG}
\title{National level particulate matter data}
\usage{
pull_national_ACAG(year)
}
\arguments{
\item{year, }{numeric object representing a selected year}
}
\value{
Dataframe object broken down by state with mean area-weighted PM2.5
values. If input year is unavailable, returns an error.
}
\description{
Pulls PM2.5 data at a national level, either internally or externally. Years
2015 through 2018 are pre-available within the package, but years ____
through 2014 are available as an external pull.
}
\examples{
pull_national_ACAG(year = 2016)
}
