% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pull_state_ACAG}
\alias{pull_state_ACAG}
\title{Pulls PM data at state level, either internally or externally}
\usage{
pull_state_ACAG(year, state, condensed = FALSE)
}
\arguments{
\item{condensed}{}
}
\value{
dataframe
}
\description{
Pulls PM data at state level, either internally or externally
}
