% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tract_ACAG.R
\name{pull_tract_ACAG}
\alias{pull_tract_ACAG}
\title{Tract level particulate matter data}
\usage{
pull_tract_ACAG(year, county_state = c())
}
\arguments{
\item{year, }{numeric object representing a selected year}

\item{county_state, }{character vector of selected counties}
}
\value{
Dataframe object broken down by census tracts in each selected county with
mean area-weighted PM2.5 values. If the county_state field is empty, PM2.5
data for census tracts in all counties is returned. If input year is
unavailable, returns an error.
}
\description{
Pulls PM2.5 data at a census tract level, either internally or as a new pull. Years
2015 through 2018 are pre-available within the package, but years 2014 and
earlier are available as a pull combining PM2.5 raster files and tigris
shape files.
}
\examples{
acag_pm_dat_tract <- pull_tract_ACAG(year = 2016, state = c("Abbeville County_SC", "Acadia Parish_LA", "Accomack County_VA"))
acag_pm_dat_tract <- pull_tract_ACAG(year = 2016)
}
