% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_county_ACAG.R
\name{save_data}
\alias{save_data}
\title{Function to save final processed data as CSV and Rdata files to output
directory}
\usage{
save_data(df, state)
}
\arguments{
\item{df}{data frame of processed data}

\item{state}{string - state abbreviation}

\item{county_name}{vector of county names (as strings) in state}
}
\description{
Function to save final processed data as CSV and Rdata files to output
directory
}
