% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_city_ACAG.R
\name{pull_city_ACAG}
\alias{pull_city_ACAG}
\title{Pulls PM data at city level, either internally or externally.
If city_state list is empty, returns condensed city list.}
\usage{
pull_city_ACAG(year, city_state = c())
}
\arguments{
\item{city_state, }{list containing desired cities in format "city_st"}
}
\value{
dataframe containing PM2.5 for tracts in each city, or PM2.5 tracts for all cities ((NOT AVAILABLE))
}
\description{
Pulls PM data at city level, either internally or externally.
If city_state list is empty, returns condensed city list.
}
