% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_county_ACAG.R
\name{pull_county_ACAG}
\alias{pull_county_ACAG}
\title{Pulls PM data at county level, either internally or externally.
If county field is empty, returns PM data for all counties.}
\usage{
pull_county_ACAG(year, county_state = c())
}
\arguments{
\item{county_state}{}
}
\value{
dataframe, for PM2.5 of counties in each state
}
\description{
Pulls PM data at county level, either internally or externally.
If county field is empty, returns PM data for all counties.
}
