% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pull_county_ACAG}
\alias{pull_county_ACAG}
\title{Pulls PM data at county level, either internally or externally}
\usage{
pull_county_ACAG(year, state, county, condensed = FALSE)
}
\arguments{
\item{condensed}{}
}
\value{
dataframe
}
\description{
Pulls PM data at county level, either internally or externally
}
