% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_national_ACAG.R
\name{pull_national_ACAG}
\alias{pull_national_ACAG}
\title{Pulls PM data at national level, either internally or externally}
\usage{
pull_national_ACAG(year)
}
\arguments{
\item{year}{}
}
\value{
dataframe with requested data
}
\description{
Pulls PM data at national level, either internally or externally
}
\examples{
pull_national_ACAG(2016)
}
