% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_state_ACAG.R
\name{pull_state_ACAG}
\alias{pull_state_ACAG}
\title{State level particulate matter data}
\usage{
pull_state_ACAG(year, state = c())
}
\arguments{
\item{year, }{numeric object representing a selected year}

\item{state, }{character vector of selected states}
}
\value{
Dataframe object broken down by counties in each state with mean
area-weighted PM2.5 values. If the state field is empty, PM2.5 data for
counties in all states is returned. If input year is unavailable, returns an
error.
}
\description{
Pulls PM2.5 data at a state level, either internally or externally. Years
2015 through 2018 are pre-available within the package, but years ____
through 2014 are available as an external pull.
}
\examples{
pull_state_ACAG(year = 2016, state = c("AL", "AK", "AZ"))
pull_state_ACAG(year = 2016)
}
