% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_state_ACAG.R, R/utilities.R
\name{pull_state_ACAG}
\alias{pull_state_ACAG}
\title{Pulls PM data at state level, either internally or externally.
If state field is empty, returns PM data for all states.}
\usage{
pull_state_ACAG(year, state, condensed = FALSE)

pull_state_ACAG(year, state, condensed = FALSE)
}
\arguments{
\item{state}{}

\item{condensed}{}
}
\value{
dataframe, for PM2.5 of counties in each state

dataframe
}
\description{
Pulls PM data at state level, either internally or externally.
If state field is empty, returns PM data for all states.

Pulls PM data at state level, either internally or externally
}
