% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_state_ACAG.R
\name{pull_state_ACAG}
\alias{pull_state_ACAG}
\title{State level particulate matter data}
\usage{
pull_state_ACAG(year)
}
\arguments{
\item{year, }{numeric object representing a selected year}
}
\value{
Dataframe object broken down by state with mean area-weighted PM2.5
values. If input year is unavailable, returns an error.
}
\description{
Pulls PM2.5 data at a state level, either internally or as a new pull. Years
2015 through 2018 are pre-available within the package, but years 2014 and
earlier are available as a pull combining PM2.5 raster files and tigris
shape files.
}
\examples{
acag_pm_dat_state <- pull_state_ACAG(year = 2016)
}
