% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tract_ACAG.R
\name{pull_tract_ACAG}
\alias{pull_tract_ACAG}
\title{Tract level particulate matter data}
\usage{
pull_tract_ACAG(
  year,
  level,
  state = c(),
  county_state = c(),
  tract_county_state = c()
)
}
\arguments{
\item{year, }{numeric object representing a selected year}

\item{level, }{character string representing desired level of pull}

\item{state, }{numeric vector of selected states via GEOID}

\item{county_state, }{numeric vector of selected counties via GEOID}

\item{tract_county_state, }{numeric vector of selected tracts via GEOID}
}
\value{
Dataframe object broken down by census tracts with
mean area-weighted PM2.5 values.
}
\description{
Pulls PM2.5 data at a census tract granularity with level selections, either
internally or as a new pull. Years 2015 through 2018 are pre-available within
the package, but years 2014 and earlier are available as a pull combining
PM2.5 raster files and tigris shape files.
}
\examples{
acag_pm_dat_tract <- pull_tract_ACAG(year = 2016, level = "National")
acag_pm_dat_tract <- pull_tract_ACAG(year = 2016, level = "State", state = c(1, 5, 4))
acag_pm_dat_tract <- pull_tract_ACAG(year = 2016, level = "County", county_state = c(45001, 22001, 51001))
acag_pm_dat_tract <- pull_tract_ACAG(year = 2016, level = "Tract", tract_county_state = c(45001950200, 45001950300, 45001950600))
}
